requirements.txt
# Django & REST API
Django==4.2.13
djangorestframework==3.15.2

# PostgreSQL
psycopg2-binary==2.9.9

# Environment
python-dotenv==1.0.1

# Machine Learning
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

# Utility
requests==2.32.3
